Add Vagrantfile and salt provisioning
This commit adds a Vagrantfile which creates a multi-machine
environment that consists of the following:

* elasticsearch:
    - This is an Ubuntu 14.04 box with Elasticsearch 1.6.
* elastalert:
    - This is an Ubuntu 14.04 box with:
        - ElastAlert and dependencies
        - Python 2.6 and 2.7
        - Supervisor

Upon instantiation, the elastalert box will be provisioned to
use elasticsearch as its Elasticsearch host. Moreover,
ElastAlert will be automatically started using the configuration
in salt/roots/salt/elastalert/config.yaml with the rules (if any)
specified in salt/roots/salt/elastalert/rules.

The automatic configuration of /etc/hosts relies on the vagrant
plugin vagrant-hostmanager, and so this plugin is required in
order to use this Vagrantfile.

Neither VirtualBox nor VMWare support hard links being created
inside shared folders, and so setup.py had to be modified in
order to prevent the creation of hard links during the run of
`python setup.py sdist` performed by tox.