Reverse nesting of term and data histograms
Reverse nesting of term and data histograms, this will open up the
possibility of pipeline aggregations in future enhancements (possibly
derivitives, moving averages etc). As the parent of a pipeline
aggregations must be a histogram or a date histogram, the date histogram
needs to be nested within the term aggregation rather than the other way
around.

Additional Minor bug fixes and  better match message.