Add match_time to alert
Currently, Kibana can only query alerts by their creation time, not the time of the event. The event time is stored in the match_body, however that is not indexed. 

This change adds the match_time field to an alert, and adds it as indexed field to the ElasticSearch index. In order to make use of the indexin feature, you will have to reindex.