Fix to ChangeRule to allow elastalert to search backwards for the last occurrences
If i have a ChangeRule type, and elastalert goes down for some time, the events that occurred during the time that elastalert was down will not be considered and there will be no alerts regarding those. And, for example, if a key had a certain value during that time, when elastalert starts again, the value is changed and it will not alert because it has no history of the values for that key.
Now, if the rule has the parameter "check_last_occurrences" and elastalert is started or the rule is modified, elastalert will fill in the occurrences variable with the last values from “time_frame” ago until now, so the next time a document with the query key appears, and the compare key is changed, an alert will be sent.