Fix Docs: Writing Filters, add ES version info
Implementing boolean logic in filters is quite different for Elasticsearch versions, mainly 2.x and 5.x. The examples given in the documentation are of Elasticsearch 2.x ( which is quite outdated ). It should be specified in documentation which version of Elasticsearch do these queries support. Users have already faced issues and confusion regarding this #912, #930