Add a changelog to keep track of version differences
I thought this might be useful to keep track of changes over time.
I am happy to maintain it, but ideally we could encourage PR authors
to add their changes as they go.

This might also be useful to get an idea of the appropriate version
change, depending on the significance of the included changes.