Fix coverage calculation to work with parallelised tests
The existing coverage.py doesn't work with xdist, which is used for
parallel testing. To fix this, use pytest-cov instead, which can
handle parallelised tests.

This brings the coverage back up to 69%, which is the same as what
coverage.py reports for the refactored files. This is a drop from
the previously reported 82%, but as far as I can tell that's entirely
due to the refactoring, not due to any files being missed or counted
incorrectly.