Fix timezone conversion check
I noticed today that `make test` had started failing on my machine
when run outside of the Docker container.

This was due to the fact that a timezone wasn't being set correctly
when run on my machine, which is on BST. The test worked fine with
the container because it was on UTC.

I traced the error down to https://github.com/jertel/elastalert2/pull/51,
in particular, a check which would return True in some cases where
it wasn't intended, because get() returns a None value by default
instead of the expected empty string.